+++ sketch/src/androidTest/java/com/github/panpf/sketch/test/fetch/ResourceUriFetcherTest.kt
@@ -18,12 +18,14 @@ package com.github.panpf.sketch.test.fetch
 import android.content.res.Resources
 import android.net.Uri
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.github.panpf.sketch.datasource.DrawableDataSource
+import com.github.panpf.sketch.datasource.ResourceDataSource
 import com.github.panpf.sketch.fetch.ResourceUriFetcher
 import com.github.panpf.sketch.fetch.newResourceUri
 import com.github.panpf.sketch.request.DisplayRequest
 import com.github.panpf.sketch.request.DownloadRequest
 import com.github.panpf.sketch.request.LoadRequest
-import com.github.panpf.sketch.test.R.drawable
+import com.github.panpf.sketch.test.R
 import com.github.panpf.sketch.test.utils.getTestContext
 import com.github.panpf.sketch.test.utils.getTestContextAndNewSketch
 import com.github.panpf.tools4j.test.ktx.assertNoThrow
@@ -101,7 +103,7 @@ class ResourceUriFetcherTest {
         val testAppPackage = context.packageName
         val fetcherFactory = ResourceUriFetcher.Factory()
         val androidResUriByName = newResourceUri(testAppPackage, "drawable", "ic_launcher")
-        val androidResUriById = newResourceUri(testAppPackage, drawable.ic_launcher)
+        val androidResUriById = newResourceUri(testAppPackage, R.drawable.ic_launcher)
         val httpUri = "http://sample.com/sample.jpg"
         val contentUri = "content://sample_app/sample"
 
@@ -155,7 +157,7 @@ class ResourceUriFetcherTest {
         }
         assertNoThrow {
             runBlocking {
-                newResourceUri(drawable.ic_launcher).let {
+                newResourceUri(R.drawable.ic_launcher).let {
                     ResourceUriFetcher(sketch, LoadRequest(context, it), Uri.parse(it))
                 }.fetch()
             }.getOrThrow()
@@ -170,7 +172,7 @@ class ResourceUriFetcherTest {
         }
         assertNoThrow {
             runBlocking {
-                newResourceUri(context.packageName, drawable.ic_launcher).let {
+                newResourceUri(context.packageName, R.drawable.ic_launcher).let {
                     ResourceUriFetcher(sketch, LoadRequest(context, it), Uri.parse(it))
                 }.fetch()
             }.getOrThrow()
@@ -185,7 +187,7 @@ class ResourceUriFetcherTest {
         }
         assertNoThrow {
             runBlocking {
-                context.newResourceUri(drawable.ic_launcher).let {
+                context.newResourceUri(R.drawable.ic_launcher).let {
                     ResourceUriFetcher(sketch, LoadRequest(context, it), Uri.parse(it))
                 }.fetch()
             }.getOrThrow()
@@ -230,5 +232,20 @@ class ResourceUriFetcherTest {
                 }.fetch()
             }.getOrThrow()
         }
+
+        Assert.assertTrue(
+            runBlocking {
+                newResourceUri(R.drawable.ic_launcher).let {
+                    ResourceUriFetcher(sketch, LoadRequest(context, it), Uri.parse(it))
+                }.fetch()
+            }.getOrThrow().dataSource is DrawableDataSource
+        )
+        Assert.assertTrue(
+            runBlocking {
+                newResourceUri(R.raw.sample).let {
+                    ResourceUriFetcher(sketch, LoadRequest(context, it), Uri.parse(it))
+                }.fetch()
+            }.getOrThrow().dataSource is ResourceDataSource
+        )
     }
 }